Implementing focus for Area
